+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][unreleased]
 
+Added:
+
+1. **cp**: `cp`
+
 Removed:
 
 1. **apt-get**: `upgrade`.