+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Added:
 
 1. **cp**: `cp`
 1. **git**: `gcb`
+1. **less**: `less`
 
 Removed: